Taken on advice from copilot as it is a genius
